backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-chess==1.999
anthropic==0.40.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
